Enhanced to be more interactive

+++ genCoverLetter.py
@@ -4,6 +4,7 @@ import shutil
 
 import openai
 from langchain.chains import ConversationalRetrievalChain, RetrievalQA
+from langchain.memory import ConversationBufferMemory
 from langchain.chat_models import ChatOpenAI
 from langchain.document_loaders import DirectoryLoader, TextLoader
 from langchain.embeddings import OpenAIEmbeddings
@@ -11,6 +12,7 @@ from langchain.indexes import VectorstoreIndexCreator
 from langchain.indexes.vectorstore import VectorStoreIndexWrapper
 from langchain.llms import OpenAI
 from langchain.vectorstores import Chroma
+from langchain.text_splitter import RecursiveCharacterTextSplitter
 
 import constants
 
@@ -20,23 +22,44 @@ os.environ["OPENAI_API_KEY"] = constants.APIKEY
 PERSONAL_DOCS_FOLDER = "personal_docs"
 PERSIST_FOLDER = PERSONAL_DOCS_FOLDER + "_persist"
 # GPT_MODEL = "gpt-3.5-turbo"
-GPT_MODEL = "gpt-3.5-turbo"
+GPT_MODEL = "gpt-3.5-turbo-16k"
+GENERATED_DOCS_FOLDER = getattr(constants, "GENERATED_DOCS_FOLDER", "~/Documents/GeneratedCoverLetters")
 
-def rebuild_embedding_chain(persist_folder, personal_docs_folder, gpt_model=GPT_MODEL):
-  print("Rebuilding the personal docs index...")
-  # if the persist_folder exists, delete it
-  if os.path.exists(persist_folder):
-    shutil.rmtree(persist_folder)
-  # create a new index
-  loader = DirectoryLoader(personal_docs_folder)
-  index = VectorstoreIndexCreator(vectorstore_kwargs={"persist_directory":PERSIST_FOLDER}).from_loaders([loader])
+chat_history = []
+
+
+def ask_conversational_question(question): 
+    answer = chain({"question": question, "chat_history": chat_history})['answer']
+    chat_history.append((question, answer))
+    return answer
+
+def get_answer(question):
+  return index.query(question)
+
+def build_embedding_chain(persist_folder, personal_docs_folder, gpt_model=GPT_MODEL, clean=False):
+  
+  if (clean):
+    print("Rebuilding the personal docs index...")
+    # if the persist_folder exists, delete it
+    if os.path.exists(persist_folder):
+      shutil.rmtree(persist_folder)
+    # create a new index
+    loader = DirectoryLoader(personal_docs_folder)
+    index = VectorstoreIndexCreator(
+      text_splitter = RecursiveCharacterTextSplitter(chunk_size=2000, chunk_overlap=100), vectorstore_kwargs={"persist_directory":PERSIST_FOLDER}).from_loaders([loader])
+  else:
+    print("Reusing vectorstore from " + PERSIST_FOLDER + " directory...\n")
+    vectorstore = Chroma(persist_directory=PERSIST_FOLDER, embedding_function=OpenAIEmbeddings())
+    index = VectorStoreIndexWrapper(vectorstore=vectorstore)
+  
   chain = ConversationalRetrievalChain.from_llm(
-    llm=ChatOpenAI(model=gpt_model),
-    retriever=index.vectorstore.as_retriever(search_kwargs={"k": 1}),
+    llm=ChatOpenAI(model=GPT_MODEL),
+    retriever=index.vectorstore.as_retriever(search_kwargs={"k": 10}),
   )
-  return chain
 
-def generate_cover_letter():
+  return (index, chain)
+
+def generate_cover_letter(applicant_name, output_folder=GENERATED_DOCS_FOLDER):
   while True:
     chat_history = []
     company_name = input("Enter the company name: ")
@@ -53,12 +76,15 @@ def generate_cover_letter():
     # Join the lines together into a String with a newline character between each line
     job_desc = "\n".join(lines)
 
-    query = "Write a job application cover letter for the person who's resume and personal details are known to you. The cover letter is for the role of " + job_title + " at company '" + company_name + "' for the following job description '" + job_desc + "'. The cover letter should be addressed to the hiring manager. The cover letter should be no more than 300 words. The cover letter should be written in a polite, friendly, and informal tone that is short, enthusiastic. Provide 3 to 5 bullet points to highlight where my skills and experience are a good for the job description, prioritizing the must have skills and experience mentioned in the above job description. Starting the letter with: \"I'm' excited to apply for the role of " + job_title + " at " + company_name + ". I believe my skills and experience are a good fit for the role:\". Add a sentence about why the company mission is important to me. Don't elaborate beyond the bullet points. Add a short closing paragraph thanking the hiring manager for their time and consideration. Sign off using the full name from the resume context"
+    most_relevant_skill = get_answer("Summarize Alex Worden's single most relevant experience to the most important requirement in following job description below using one or two words to state the type of experience. \n" + job_desc)
 
-    print("\n\n... Generating cover letter ...\n\n")
-    result = chain({"question": query, "chat_history": chat_history})
-    print("Cover Letter:\n")
-    print(result['answer'])
+    print("\nAlex Worden's most relevant experience to the most important requirement in the job description is: " + most_relevant_skill + "\n")
+
+    query = "Write a job application cover letter for me, " + applicant_name + ". The cover letter is for the role of " + job_title + " at company '" + company_name + "' for the following job description '" + job_desc + "'. The cover letter should be addressed to the hiring manager. The cover letter should be no more than 400 words. The cover letter should be written in a polite, friendly, and informal tone that is genuine and enthusiastic. Start the letter with: \"I'm excited to apply for the role of " + job_title + " at " + company_name + ". I believe the following skills and experience I have are a good fit for the role:\". Provide only 3 to 5 bullet points that highlight my skills and experience that are a good match for the requirements stated in the job description. Do not directly quote the phrases used in the job description, but instead, quote my relevant skills and experience. Add a genuine, but not overstated sentence about why the company mission is important to me. Don't elaborate too much. Add a short closing paragraph thanking the hiring manager for their time and consideration."
+
+    print("\n\n... Generating cover letter ...\n ")
+    cover_letter = ask_conversational_question(query)
+    print("Cover Letter:\n\n" + cover_letter + "\n")
     print("\n========================\n")
 
     #Generate a PDF of the cover letter and save to a file named with the company name and job title
@@ -66,13 +92,13 @@ def generate_cover_letter():
     pdf = FPDF()
     pdf.add_page()
     pdf.set_font("Arial", size=12)
-    pdf.multi_cell(0, 5, result['answer'])
-    pdf.output(company_name + "_" + job_title + ".pdf")
-
-    #ask the user if they want to generate another cover letter
-    # if they say no, exit the program
-    # if they say yes, repeat the process
-    # if they enter an invalid response, ask them again
+    pdf.multi_cell(0, 5, cover_letter)
+    # write the pdf to the output folder
+    if not os.path.exists(output_folder):
+      os.makedirs(output_folder)
+    pdf.output(output_folder + "\\" + company_name + "_" + job_title + ".pdf")
+    
+    # ask the user if they want to generate another cover letter
     while True:
       answer = input("Generate another cover letter? (y/n): ")
       if answer == "n":
@@ -84,37 +110,40 @@ def generate_cover_letter():
 
 
 # ================== START OF MAIN PROGRAM ==================
-# If PERSIST is enabled and the persist directory exists, reuse the index
-if os.path.exists(PERSIST_FOLDER): 
-  print("Reusing vectorstore from " + PERSIST_FOLDER + " directory...\n")
-  vectorstore = Chroma(persist_directory=PERSIST_FOLDER, embedding_function=OpenAIEmbeddings())
-  index = VectorStoreIndexWrapper(vectorstore=vectorstore)
-  chain = ConversationalRetrievalChain.from_llm(
-    llm=ChatOpenAI(model=GPT_MODEL),
-    retriever=index.vectorstore.as_retriever(search_kwargs={"k": 1}),
-  )
-else:
-  chain = rebuild_embedding_chain(PERSIST_FOLDER, PERSONAL_DOCS_FOLDER)
 
-chat_history = []
+# if os.path.exists(PERSIST_FOLDER): 
+#   chain = build_embedding_chain(PERSIST_FOLDER, PERSONAL_DOCS_FOLDER, clean=False)
+# else:
+(index, chain) = build_embedding_chain(PERSIST_FOLDER, PERSONAL_DOCS_FOLDER, clean=True)
 
-print("Ask questions about your skills and experience from documents in the " + PERSIST_FOLDER + " folder. Enter 'quit' to exit.\n")
+# result = chain({"question": "Tell me the name of the person that the resume is about? Reply with only the name and use the full name if you have that information, otherwise reply with the token UNKNOWN", "chat_history": chat_history})
+# applicant_name = result['answer']
+
+applicant_name = get_answer("What is the name of the person that the resume is about? Reply with only the name and use the full name if you have that information, otherwise reply with the token UNKNOWN")
+
+companies_worked_for = get_answer("List all of the companies that " + applicant_name + " has worked for in a bullet list in reverse chronological order with start and end dates")
+
+print(applicant_name + " has worked for the following companies:\n" + companies_worked_for + "\n")
+
 while True:
-  query = input(": ")
-  if query in ['quit', 'q', 'exit']:
+  print("=====================================================\n" +
+        "Ask a question about " + applicant_name + ".\n" + 
+        "Enter 'q' to exit.\n" +
+        "Enter 'Gen Cover Letter' to generate a cover letter.\n" +
+        "=====================================================\n")
+  userInput = input("Q: ")
+  if userInput in ['quit', 'q']:
     break
-  elif query == "Gen Cover Letter":
-    generate_cover_letter()
-  elif query == "Refresh":
-    chain = rebuild_embedding_chain(PERSIST_FOLDER, PERSONAL_DOCS_FOLDER)
+  elif userInput == "Gen Cover Letter":
+    generate_cover_letter(applicant_name="Alex Worden")
+    print("Ask more questions about the skills and experience of " + applicant_name + ". Enter 'quit' to exit.\n")
+  elif userInput == "Refresh":
+    index = build_embedding_chain(PERSIST_FOLDER, PERSONAL_DOCS_FOLDER)
   else:
-    result = chain({"question": query, "chat_history": chat_history})
-    print(result['answer'])
+    userInput = "answer the following question '" + userInput + "' in relation to " + applicant_name + ". Assume that references to 'I' and 'me' refer to " + applicant_name + ". You answers should substitute the personal pronouns 'I' and 'me' with the name " + applicant_name + "."
+    print("\n... Generating answer ...\n")
+    answer = ask_conversational_question(userInput)
+    print(answer + "\n")
 
-    chat_history.append((query, result['answer']))
-    query = None
+    #chat_history.append((userInput, result['answer']))
 
-# BEGIN: 9f8c6549bwf9
-
-# END: 9f8c6549bwf9
-
